requirements.txt
langchain==0.1.5
databricks-vectorsearch==0.40
sqlalchemy==2.0.32
openai==1.40.1
mlflow==2.13.1